Refactored more frontend code
