--- web/components/SceneSwitchButton.js
@@ -1,36 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-
-import { updateCamera } from '../actions/spaceSceneActions';
-
-class SceneSwitchButton extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = { viewingMoon: false };
-
-    this.handleToggle = this.handleToggle.bind(this);
-  }
-
-
-  handleToggle() {
-    this.setState({ viewingMoon: !this.state.viewingMoon });
-    this.state.viewingMoon ? this.props.updateCamera('solar')
-      : this.props.updateCamera('moon');
-  }
-
-  render() {
-    return(
-      <div className="changeSceneButton">
-        <button onClick={this.handleToggle}>{
-          this.state.viewingMoon ? 'View Earth' : 'View Moon'
-          }
-        </button>
-      </div>
-    )
-  }
-}
-
-const mapStateToProps = state => state;
-
-export default connect(mapStateToProps, { updateCamera })(SceneSwitchButton)
