+++ web/components/sliders/TimelineSlider.js
@@ -4,9 +4,9 @@ import { Slider, Rail, Handles, Tracks, Ticks } from "react-compound-slider";
 import { SliderRail, Handle, Track, Tick } from "./sliderComponents"; // example render components - source in sliderComponents
 import { format, getTime } from "date-fns";
 import { scaleTime } from "d3-scale";
-import { updateSimulationTime } from '../actions/spaceSceneActions';
-import { selectSimulationTime } from '../reducers';
-import { set_working_date, get_coverage, get_working_date } from '../libraries/position_store';
+import { updateSimulationTime } from '../../actions/spaceSceneActions';
+import { selectSimulationTime } from '../../reducers';
+import { set_working_date, get_coverage } from '../../libraries/position_store';
 
 const sliderStyle = {
   position: "relative",
