+++ web/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { render } from "react-dom";
 import SpaceScene from './components/SpaceScene';
-import SceneSwitchButton from './components/SceneSwitchButton';
 import './styles/styles.scss';
 import SimulationControls from "./components/SimulationControls";
 import { Provider } from 'react-redux';
